Updated commands, test cases and user guide.
The /add command only handles 2 fields: p/ and e/.
There is also no command to sort the addressbook.

Since Week 2 LOs require the AddressBook to support an additional field to the /add command and a /sort command, this has raised an inconsistency.

Thus, a dob/ field has been added to the /add command, and a /sort command has been added.
With the above addition, the test cases have been improved as well to include 1) testing empty input, and 2) testing /add with the dob/ field.

The following have been updated in the user guide:
/add now supports the dob/ prefix.
/sort has been added that sorts the addressbook in ASC order by name.